+++ Sources/KeyboardKitSwiftUI/System/SystemKeyboardStyle.swift
@@ -14,9 +14,9 @@ import SwiftUI
  This struct provides static properties that are used by the
  various extensions in `View+System`.
  
- `TODO` This is not a viable way forward, since first of all
- this is layout, not style. Also, layout is a lot harder and
- cannot be specified with single values.
+ `TODO` The class has layout properties that should and will
+ be removed in future versions, as soon as they are not used
+ by the library. Be careful with using this until `4.0`.
  */
 public class SystemKeyboardStyle: ObservableObject {
     
@@ -26,18 +26,12 @@ public class SystemKeyboardStyle: ObservableObject {
         buttonCornerRadius: CGFloat,
         buttonHeight: CGFloat = .standardKeyboardRowHeight(),
         buttonInsets: EdgeInsets = .insets(from: .standardKeyboardRowItemInsets()),
-        buttonSpacing: CGFloat,
-        padding: EdgeInsets,
-        rowSpacing: CGFloat,
         shiftWidth: CGFloat) {
         self.backspaceWidth = backspaceWidth
         self.bottomRowSpacePercentage = bottomRowSpacePercentage
         self.buttonCornerRadius = buttonCornerRadius
         self.buttonHeight = buttonHeight
         self.buttonInsets = buttonInsets
-        self.buttonSpacing = buttonSpacing
-        self.padding = padding
-        self.rowSpacing = rowSpacing
         self.shiftWidth = shiftWidth
     }
     
@@ -45,10 +39,7 @@ public class SystemKeyboardStyle: ObservableObject {
     public let bottomRowSpacePercentage: CGFloat
     public let buttonHeight: CGFloat
     public let buttonInsets: EdgeInsets
-    public let buttonSpacing: CGFloat
     public let buttonCornerRadius: CGFloat
-    public let padding: EdgeInsets
-    public let rowSpacing: CGFloat
     public let shiftWidth: CGFloat
 }
 
@@ -63,9 +54,6 @@ public extension SystemKeyboardStyle {
             backspaceWidth: 50.0,
             bottomRowSpacePercentage: 0.5,
             buttonCornerRadius: 4.0,
-            buttonSpacing: 6.0,
-            padding: EdgeInsets(top: 4, leading: 4, bottom: 4, trailing: 4),
-            rowSpacing: 12.0,
             shiftWidth: 50.0)
     }
 }